requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-limiter
psycopg2-binary==2.9.7
argon2-cffi